# Gex.NetCore/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Gex",
    "description": "Colleción de API's correspondientes al sistema de exámenes Gex.",
    "contact": {
      "name": "Agustin Ibañez",
      "email": "admin@gexsystem.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/examen": {
      "get": {
        "tags": [
          "Examen"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseICollectionGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseICollectionGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseICollectionGexResult"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Examen"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamenRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExamenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Examen"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamenRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamenRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ExamenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Examen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/examen/{id}/preguntas": {
      "get": {
        "tags": [
          "Examen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/examen/{id}": {
      "get": {
        "tags": [
          "Examen"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamenResponseGexResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Object": {
        "type": "object",
        "nullable": true
      },
      "GexErrorMessage": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8,
          9,
          10
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/Object"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": {
          "$ref": "#/components/schemas/Object"
        },
        "nullable": true
      },
...
      "MateriaTipo": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "MateriaResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "tipo": {
            "$ref": "#/components/schemas/MateriaTipo"
          }
        },
        "nullable": true
      },
      "ExamenTipo": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "ExamenResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "materia": {
            "$ref": "#/components/schemas/MateriaResponse"
          },
          "tipo": {
            "$ref": "#/components/schemas/ExamenTipo"
          },
          "nota_regular": {
            "type": "integer",
            "format": "int32"
          },
          "nota_promo": {
            "type": "integer",
            "format": "int32"
          },
          "recuperatorio": {
            "type": "boolean"
          }
        },
        "nullable": true
      },
      "ExamenResponseICollectionGexResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamenResponse"
            },
            "nullable": true
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "nullable": true
              },
              "nullable": true
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/GexErrorMessage"
          }
        },
        "nullable": true
      },
      "ExamenRequest": {
        "required": [
          "id",
          "materia_id",
          "nota_promo",
          "nota_regular",
          "tipo"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tipo": {
            "$ref": "#/components/schemas/ExamenTipo"
          },
          "nota_regular": {
            "maximum": 10,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "nota_promo": {
            "maximum": 10,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "recuperatorio": {
            "type": "boolean"
          },
          "materia_id": {
            "type": "integer",
            "format": "int64"
          }
        },
        "nullable": true
      },
      "ExamenResponseGexResult": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExamenResponse"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "error_messages": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string",
                "nullable": true
              },
              "nullable": true
            },
            "nullable": true
          },
          "error_code": {
            "$ref": "#/components/schemas/GexErrorMessage"
          }
        },
        "nullable": true
      },
...
    }
  }
}